Fixed bug with token verify, wired StarService
There was a latent issue with token verification where I was
JSON.stringify'ing too many times, leading to invalid tokens when
actually they should've validated.

Database should allow for models.repr to be NULL for as-of-yet
uninitialized models. We should assume that when repr is non-NULL, then
the string inside it is a valid serialized graph. Our application should
make sure to maintain this invariant, which is why I eliminated the
empty string default.

StarService is now wired in and listening and /star.